// frontend/src/utils/validation-utils.d.ts
import { ValidationType, OptionsType } from "../types/validation.type";
export declare class ValidationUtils {
    static validateField(element: HTMLElement | null, options: OptionsType | undefined): boolean;
    static validateForm(validations: ValidationType[]): boolean;
}
//# sourceMappingURL=validation-utils.d.ts.map